Update App.css
There is an issue with the About page's description text pushing the social links past the About panel background. It appears to occur on Google Chrome mobile portrait view.

When in landscape view, the description text is properly scrollable and fits within the panel background.

I do not have a mobile emulator and the Developer Tools in my browsers are unable to replicate the bug seen via my mobile.

So I've made an adjustment to the CSS, using vw instead of % to see if that amends the issue, and this is the deployment of that change. Will check mobile rendering after deployment. Fingers crossed.