Adding sound effects o menu buttons and project links
It turns out the bug I discovered whereby the About description pushes buttons outside of their background div in Chrome portrait view, is most likely having to do with that mobile phone itself.

Checking another mobile phone, the issue does not happen at all.

So instead of looking into that route, I added sound effects to the footer menu buttons and to the project link buttons.

I was thinking of adding more, but refrained as I didn't want to oversaturate the site with sound such that it got too annoying to listen to.