Added button hover effect to make icons smaller
This was a lot simpler than I expected. Initially I thought the CSS code would have to update the sizes for every sub element of each button, but Bing Chat once again saved the day with a very simple attribute for the container element, transform: scale(0.85). Learning as you go is fun.