Add scroll arrow label and rename Twitter button
I added a combined scroll and arrow label at the bottom of the About description. It displays when the user has not yet scrolled to the bottom of the description text. When they do, then the label disappears, as intended.

I also renamed the button from Twitter to X (Twitter) given their current rebranding effort.