Fixed page scroll bug
The React version I am using seems to inherently remember where vertically I am scrolled down in the Blog page, and when I click a link to a post, it auto scrolls that far down the post, which is inconvenient.

Thanks to Bing chat, the code has been amended such that it auto-scrolls to the top of the Post's page when a link from the Blog page is clicked via componentDidMount() and scrollTo().