Changing design for code tag blocks
In my previous site designs, I made it so that <code> blocks would scroll horizontally with a scrollbar.

In the current design I removed that ability and just wrapped the text. This is not a useful, so I reverted the design so that code content is more easily read and copy-and-pasted.