Adding a loading screen for main page of site
I noticed the portfolio page often takes a notable amount of time to render the gifs used as project preview images onscreen.

The purpose of this loading screen is to pre-load those assets before access to the main page is provided. I have only applied this for the main home page as a live "test", because testing locally doesn't reflect the time taken to download the images across the net.