2nd attempt at implementing a loading screen
A loading screen is intended to appear when a user first logs into the site. It preloads all site assets, so that when you navigate to other pages, the images are ready. Currently the user has to wait a notable amount of time once already on that page, which isn't aesthetically right.

I cannot test this loading screen locally. I am testing it on mobile because I can easily clear cache there for each test.

This refactor involved affecting most React components, as I had to change the methodology of accessing assets using a destructure method.