Testing mobile browser app
The previous update I did to add animated squares seems to work fine on desktop browser.

But when I check my mobile browser app, the height of the page seems to have doubled with empty space in the latter half.

I suspect this has to do with the height and width definitions added to .bg-home.

I have commented those out temporarily to see if that fixes the issue.