Testing fix for mobile app
The About/Home page is experiencing a glitch where the menu bars seems to be moving downward. Th issue only appears on mobile browsers, not desktop browsers.

I tried removing the width and height explicit definition on bg-home to no avail.

I've restored that. I also noticed that .bg-home-background  is missing position: absolute which the other elements have, so I am giving that a try.