Fix to display floating squares
This was a fascinating fix for me.

In order to have squares float on the screen, they needed to be appended to a div which had a defined height and width. ChatGPT asked me to apply that height and width to the div bg-home, which I specified I didn't want to do because it would upset the rest of the layout.

After a couple failed tries with ChatGPT to solve the problem, I solved the problem myself, realising that I could just append the squares to a different div that already did have its height and width defined.