Refactored blog post markdown files and added tag filter to Portfolio and Blog pages
2 major changes.

1. Refactored blog post markdown file header.
Previously, tech attribute was a generic name field, and tag attribute was a single value field linked to a logo.
Now, tech attribute links to a logo, and tag attribute has been changed to tags attribute which is an array.

Associated Post-related components and Main.js post parser are updated to reflect this refactoring.

2. Added filter to Portfolio Previews and Blog Post Previews components. The filter extracts the tags from projects and blog posts, and displays the unique values as a filter, with the first button being an "ALL" filter.